Add react router library, and started generating routes for the app
